requirements.txt
asgiref==3.2.4
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-adminlte-ui==1.5.0
django-adminlte3==0.1.2
django-bootstrap4==1.1.1
django-jet==1.0.8
django-robots==4.0
django-treebeard==4.3
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
selenium==3.141.0
